# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  app.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: API documentation for KubeCloud.
  title: KubeCloud API
  version: "1.0"
paths:
...
  /nodes:
    get:
      consumes:
      - application/json
      description: Retrieves a list of nodes from the grid proxy based on the provided
        filters.
      operationId: list-nodes
      parameters:
      - description: 'Filter by healthy nodes (default: true)'
        in: query
        name: healthy
        type: boolean
      - description: 'Filter by rentable nodes (default: true)'
        in: query
        name: rentable
        type: boolean
      - description: 'Limit the number of nodes returned (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Nodes are retrieved successfully
          schema:
            $ref: '#/definitions/app.APIResponse'
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      summary: List nodes
      tags:
      - nodes
...
  /user/nodes:
    get:
      consumes:
      - application/json
      description: Returns a list of reserved nodes for a user
      operationId: list-reserved-nodes
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/app.APIResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: List reserved nodes
      tags:
      - nodes
  /user/nodes/{node_id}:
    post:
      consumes:
      - application/json
      description: Reserves a node for a user
      operationId: reserve-node
      parameters:
      - description: Node ID
        in: path
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Reserve node
      tags:
      - nodes
  /user/nodes/unreserve/{contract_id}:
    delete:
      consumes:
      - application/json
      description: Unreserve a node for a user
      operationId: unreserve-node
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Unreserve node
      tags:
      - nodes
...
swagger: "2.0"